moving create-member into members API
