+++ pages/edit/index.tsx
@@ -1,4 +1,4 @@
-import Button, { ButtonSize } from "@/components/Button";
+import Button from "@/components/Button";
 import { Heading, Subheading } from "@/components/Heading";
 import MetaTags from "@/components/Metatags";
 import Nav from "@/components/Nav";
@@ -6,10 +6,19 @@ import Plausible from "@/components/Plausible";
 import Input from "@/components/form/Input";
 import Label from "@/components/form/Label";
 import Head from "next/head";
-import { useState } from "react";
-import { sendSignInLinkToEmail } from "firebase/auth";
+import { useEffect, useState } from "react";
+import {
+  isSignInWithEmailLink,
+  sendSignInLinkToEmail,
+  signInWithEmailLink,
+} from "firebase/auth";
 import { auth } from "@/lib/firebase";
 import Code from "@/components/Code";
+import { useRouter } from "next/router";
+import LoadingSpinner, {
+  LoadingSpinnerVariant,
+} from "@/components/LoadingSpinner";
+import Link from "next/link";
 
 export async function getServerSideProps(context) {
   const { req } = context;
@@ -34,20 +43,31 @@ export default function EditPage({ baseUrl, pageTitle }) {
       </Head>
       <Nav backUrl="/" />
       <Heading>Welcome back, Hawaiian.</Heading>
-      <Subheading centered>
-        First, let's sign you in with the email you registered with.
-      </Subheading>
       <RequestForm baseUrl={baseUrl} />
     </>
   );
 }
 
+enum PageState {
+  Loading = "LOADING",
+  EmailSent = "EMAIL_SENT",
+  NotLoggedIn = "NOT_LOGGED_IN",
+  Error = "ERROR",
+}
+
 function RequestForm({ baseUrl }) {
+  const router = useRouter();
   const [email, setEmail] = useState<string>("");
-  const [emailSent, setEmailSent] = useState<boolean>(false);
+  const [pageState, setPageState] = useState<PageState>(PageState.NotLoggedIn);
+  const [errorReason, setErrorReason] = useState<string>("");
+  const [errorBackup, setErrorBackup] = useState<string>("");
+
+  const backToLogin = () => {
+    setPageState(PageState.NotLoggedIn);
+  };
 
   const handleSignIn = () => {
-    const fullUrl = `${baseUrl}/edit/finish-login`;
+    const fullUrl = `${baseUrl}/edit`;
     const actionCodeSettings = {
       url: fullUrl,
       handleCodeInApp: true,
@@ -61,8 +81,66 @@ function RequestForm({ baseUrl }) {
         const errorMessage = error.message;
         console.error("error sending email:", errorCode, errorMessage);
       });
-    setEmailSent(true);
+    setPageState(PageState.EmailSent);
   };
+
+  const fetchMemberMapping = async (token: string) => {
+    const response = await fetch("/api/member-id", {
+      method: "GET",
+      headers: {
+        "Content-Type": "application/json",
+        Authorization: `Bearer ${token}`,
+      },
+    });
+    if (!response.ok) {
+      if (response.status === 404) {
+        setErrorReason("we don't have this email with any current members.");
+        setErrorBackup("Can't remember which email you used?");
+      }
+      setPageState(PageState.Error);
+    }
+    const data = await response.json();
+    const memberId = data.memberId;
+    router.push({
+      pathname: `/edit/member/`,
+      query: { memberId },
+    });
+  };
+
+  useEffect(() => {
+    if (isSignInWithEmailLink(auth, window.location.href)) {
+      setPageState(PageState.Loading);
+      let email = window.localStorage.getItem("emailForSignIn");
+      if (!email) {
+        // User opened the link on a different device. To prevent session fixation
+        // attacks, ask the user to provide the associated email again:
+        email = window.prompt(
+          "Looks like you've opened the sign-in link in a different " +
+            "window or device. For security reasons, please enter your " +
+            "email again for confirmation.",
+        );
+        window.localStorage.setItem("emailForSignIn", email);
+      }
+      signInWithEmailLink(auth, email, window.location.href)
+        .then((result) => {
+          if (result.user.email !== email) {
+            setErrorReason("the email you signed in with doesn't match.");
+            setPageState(PageState.Error);
+          } else {
+            window.localStorage.removeItem("emailForSignIn");
+            result.user.getIdToken().then((idToken) => {
+              fetchMemberMapping(idToken);
+            });
+          }
+        })
+        .catch((error) => {
+          if (error.code === "auth/invalid-action-code") {
+            setErrorReason("the link is invalid or expired.");
+          }
+          setPageState(PageState.Error);
+        });
+    }
+  }, []);
 
   return (
     <div
@@ -77,16 +155,11 @@ function RequestForm({ baseUrl }) {
         px-4
       `}
     >
-      {emailSent ? (
-        <div className="text-center mt-4">
-          <h2 className="text-base">
-            You should have <em>just</em> received a sign in email from us. If
-            you didn't, you may need to add{" "}
-            <Code>no-reply@hawaiiansintech.org</Code> to your address book.
-          </h2>
-        </div>
-      ) : (
+      {pageState === PageState.NotLoggedIn ? (
         <>
+          <Subheading centered>
+            First, let's sign you in with the email you registered with.
+          </Subheading>
           <div className="mb-8 w-3/4">
             <div className="mb-2">
               <Label label={"Email:"} />
@@ -100,6 +173,55 @@ function RequestForm({ baseUrl }) {
             />
           </div>
           <Button onClick={handleSignIn}>Sign In</Button>
+        </>
+      ) : pageState === PageState.EmailSent ? (
+        <div className="text-center mt-4">
+          <h2 className="text-base">
+            You should receive a sign in email from us real soon. If you didn't,
+            you may need to add <Code>no-reply@hawaiiansintech.org</Code> to
+            your address book.
+          </h2>
+        </div>
+      ) : pageState === PageState.Loading ? (
+        <>
+          <Subheading centered> Logging in...</Subheading>
+          <div className="flex w-full justify-center p-4">
+            <LoadingSpinner
+              className="w-16 h-16"
+              variant={LoadingSpinnerVariant.Invert}
+            />
+          </div>
+        </>
+      ) : (
+        <>
+          <div className="text-center mt-4">
+            <p className="text-2xl mb-8">
+              Gonfunnit, looks like{" "}
+              {errorReason ? errorReason : "something went wrong"}
+            </p>
+            <div className="mb-8">
+              <Button onClick={backToLogin}>
+                {errorReason === "the link is invalid or expired."
+                  ? "Try logging in again"
+                  : "Try another email"}
+              </Button>
+            </div>
+            <p>
+              {errorBackup
+                ? errorBackup
+                : "If you still having trouble, no worries"}
+              <br />
+              Contact{" "}
+              <Link href="mailto:kekai@hawaiiansintech.org" target="_blank">
+                kekai
+              </Link>{" "}
+              or{" "}
+              <Link href="mailto:kamakani@hawaiiansintech.org" target="_blank">
+                kamakani
+              </Link>{" "}
+              and we&rsquo;ll get you sorted out.
+            </p>
+          </div>
         </>
       )}
     </div>
